+++ src/main/java/com/nlmk/kb/server/repository/ResultsConfigRepository.java
@@ -5,8 +5,11 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface ResultsConfigRepository extends JpaRepository<ResultsConfig, Long> {
+public interface ResultsConfigRepository extends JpaRepository<ResultsConfig, Integer> {
 
+    /**
+     * Поиск конфигурации по наименованию головного объекта AVRO схемы
+     */
     List<ResultsConfig> findByAvroName(String avroName);
 
 }
